refactor(crm-api): extract aggregation engines + messaging workers
buildAggregationEngines (wire_aggregation.go) moves the messages+gchat
engine construction, gchat rematch gate, and reenqueuer-registry .set;
registerMessagingWorkers (wire_scheduler.go) moves the aggregate+sweeper
worker registration + periodic job (via reg.addPeriodic). telegram
manager / gchat provider threaded as nil-safe params.
