+++ backend/cmd/crm-api/main.go
@@ -37,20 +37,15 @@ import (
 	"personal-crm/backend/internal/api/handlers"
 	"personal-crm/backend/internal/auth"
 	"personal-crm/backend/internal/config"
-	"personal-crm/backend/internal/consumer"
-	"personal-crm/backend/internal/consumer/consumerjobs"
 	"personal-crm/backend/internal/db"
 	"personal-crm/backend/internal/events"
-	"personal-crm/backend/internal/google"
 	"personal-crm/backend/internal/health"
 	"personal-crm/backend/internal/logger"
-	"personal-crm/backend/internal/messages"
 	"personal-crm/backend/internal/repository"
 	"personal-crm/backend/internal/scheduler"
 	"personal-crm/backend/internal/service"
 
 	"github.com/gin-gonic/gin"
-	"github.com/google/uuid"
 	"github.com/riverqueue/river"
 	"github.com/riverqueue/river/riverdriver/riverpgxv5"
 	swaggerFiles "github.com/swaggo/files"
@@ -188,12 +183,10 @@ func run() int {
 
 	// Message-store repos + staging registry + venue resolver.
 	messaging := buildMessagingFoundation(database.Queries, messagesMessageRepo, calendarRepoForIngest)
-	commsMessageRepo := messaging.CommsMessageRepo
 
 	// Event-bus consumers (Cadence / Knowledge / FollowUp / InteractionRecorder)
 	// + their shared collaborators, wired into ContactService via its setters.
 	consumers := buildEventConsumers(cfg, database, core, graph, ingest, messaging, eventBus, riverClient)
-	aggregatorReenqueuerHolder := consumers.AggregatorReenqueuerHolder
 
 	// IngestService + meeting-note conflict-resolution surface. Hoisted
 	// after the consumers so the call.* inline handler can reuse them in
@@ -269,137 +262,11 @@ func run() int {
 		importHandler.SetPostImportHook(telegramManager)
 	}
 
-	// Messages aggregator engine + reenqueuer + worker + sweeper.
-	// Wired unconditionally — the Mac daemon push pipeline accepts
-	// raw_message.* envelopes regardless of any feature flag, and the
-	// engine is a stateless function over messagesMessageRepo (no
-	// daemon-side connection or background loop).
-	//
-	// The chat-aware AggregateForContact path is what preserves the
-	// engine's extend/bridge/coalesce contract. The
-	// MessagingAggregateForContactWorker iterates over the contact's
-	// distinct unprocessed chats and invokes it per chat; the periodic
-	// sweeper provides a 5-min safety net for the never-claimed
-	// stranded-row gap.
-	messagesEnqueuer := consumer.NewRiverInteractionRecorderEnqueuer(riverClient)
-	const messagesBurstWindowHours = 4
-	const messagesReplyBridgeHours = 48
-	messagesEngine := messages.NewAggregationEngine(
-		messagesBurstWindowHours,
-		messagesReplyBridgeHours,
-		messagesMessageRepo,
-		interactionRepo,
-		contactService,
-		contactService,
-		eventBus,
-		database.Pool,
-		messagesEnqueuer,
-	)
-	messagesReenqueuer := consumer.NewMessagesAggregatorReenqueuer(
-		messagesEngine,
-		riverClient,
-		repository.InteractionSourceMessages,
-	)
+	// Aggregation engines (messages + gchat) + reenqueuer registry.
+	agg := buildAggregationEngines(database, core, graph, ingest, messaging, consumers, eventBus, riverClient, telegramManager, gchatProvider, gchatSyncStates)
 
-	// GChat aggregation engine over comms_message. LIVE but INERT: the
-	// engine/worker/sweeper/reenqueuer for gchat run on every tick, but every
-	// query is source='gchat'-scoped and returns zero rows until a provider +
-	// enablement write comms_message(source='gchat') rows. Burst/reply windows
-	// are hard-coded here (matching how messages hard-codes its constants);
-	// env-var overrides are out of scope for now.
-	gchatEnqueuer := consumer.NewRiverInteractionRecorderEnqueuer(riverClient)
-	const gchatBurstWindowHours = 2
-	const gchatReplyBridgeHours = 48
-	gchatEngine := google.NewGChatAggregationEngine(
-		gchatBurstWindowHours,
-		gchatReplyBridgeHours,
-		commsMessageRepo,
-		interactionRepo,
-		contactService,
-		contactService,
-		eventBus,
-		database.Pool,
-		gchatEnqueuer,
-	)
-
-	// GChat rematch handlers: registered when the provider was constructed
-	// (Google OAuth configured) AND external sync is enabled. They are provably
-	// inert until an enabled gchat sync state exists — each gates FIRST on
-	// ListEnabledSyncStates filtered to source='gchat' and returns (0, nil) when
-	// that set is empty. The email handler co-registers under "email" alongside
-	// Gmail/Calendar; its gchat-scoped gate means it no-ops while the others do
-	// their real work. The provider itself is NOT registered into
-	// providerRegistry (the scheduler never runs it).
-	if gchatProvider != nil && gchatSyncStates != nil {
-		rematchService.Register(google.NewGChatHandleRematchHandler(gchatProvider, gchatSyncStates, commsMessageRepo, gchatEngine))
-		rematchService.Register(google.NewGChatEmailRematchHandler(gchatProvider, gchatSyncStates, commsMessageRepo, gchatEngine))
-		logger.Info().Msg("GChat rematch handlers registered (inert until a gchat sync state is enabled)")
-	}
-
-	gchatReenqueuer := consumer.NewCommsAggregatorReenqueuer(
-		gchatEngine,
-		riverClient,
-		repository.InteractionSourceGChat,
-	)
-
-	// Wire the per-source aggregator reenqueuer registry. The
-	// InteractionRecorderWorker holds the deferred holder; this
-	// assignment makes the post-commit reenqueue path live for both
-	// telegram-source and messages-source events. When Telegram is
-	// disabled the telegram entry is a no-op reenqueuer (so calls for
-	// telegram-source envelopes — which won't be produced anyway —
-	// degrade cleanly).
-	reenqueuerEntries := map[string]consumer.AggregatorReenqueuer{
-		repository.InteractionSourceMessages: messagesReenqueuer,
-		repository.InteractionSourceGChat:    gchatReenqueuer,
-	}
-	if telegramManager != nil {
-		reenqueuerEntries[repository.InteractionSourceTelegram] = consumer.NewTelegramAggregatorReenqueuer(telegramManager.AggregationEngine())
-	} else {
-		reenqueuerEntries[repository.InteractionSourceTelegram] = consumer.NoopAggregatorReenqueuer{}
-	}
-	aggregatorReenqueuerHolder.set(consumer.NewAggregatorReenqueuerRegistry(reenqueuerEntries))
-
-	// Register the messaging aggregate workers. The chat-lister
-	// registry maps source → repository's ListUnprocessedChatsByContact;
-	// future messaging sources (whatsapp etc) extend the map without
-	// touching the worker.
-	chatListerRegistry := scheduler.NewPerSourceChatListerRegistry(
-		map[string]func(ctx context.Context, contactID uuid.UUID) ([]string, error){
-			repository.InteractionSourceMessages: messagesMessageRepo.ListUnprocessedChatsByContact,
-			// Source-bound closure: the comms repo method is multi-source
-			// (ListUnprocessedChatsByContactForSource), so bind 'gchat'.
-			repository.InteractionSourceGChat: func(ctx context.Context, contactID uuid.UUID) ([]string, error) {
-				return commsMessageRepo.ListUnprocessedChatsByContactForSource(ctx, repository.InteractionSourceGChat, contactID)
-			},
-		},
-	)
-	river.AddWorker(riverWorkers, scheduler.NewMessagingAggregateForContactWorker(
-		map[string]scheduler.ChatAwareAggregator{
-			repository.InteractionSourceMessages: messagesEngine,
-			repository.InteractionSourceGChat:    gchatEngine,
-		},
-		chatListerRegistry,
-	))
-
-	// Periodic 5-min sweeper — drains never-claimed stranded rows that
-	// the in-line worker re-list loop AND the post-Stage-3 reenqueue
-	// both missed. Run once on startup so restart-recovery does not wait
-	// a full interval before the safety net engages.
-	sweeperListers := map[string]scheduler.UnprocessedContactLister{
-		repository.InteractionSourceMessages: messagesMessageRepo,
-		// Source-bound adapter: comms_message is multi-source, so wrap the
-		// repo with a 'gchat'-pinned lister.
-		repository.InteractionSourceGChat: newCommsSourceContactLister(commsMessageRepo, repository.InteractionSourceGChat),
-	}
-	river.AddWorker(riverWorkers, scheduler.NewMessagingAggregateSweeperWorker(sweeperListers, riverClient))
-	riverClient.PeriodicJobs().Add(river.NewPeriodicJob(
-		river.PeriodicInterval(5*time.Minute),
-		func() (river.JobArgs, *river.InsertOpts) {
-			return consumerjobs.MessagingAggregateSweeperArgs{}, nil
-		},
-		&river.PeriodicJobOpts{RunOnStart: true},
-	))
+	// Register the chat-aware messaging aggregate worker + sweeper (+ periodic).
+	registerMessagingWorkers(reg, ingest, messaging, agg, riverClient)
 
 	// Initialize handlers
 	contactHandler := handlers.NewContactHandler(contactService)
